requirements.txt
PyQt5==5.15.10
Pillow==10.1.0
pyqtdarktheme==2.1.0
